融合Server与FakeContext
