+++ server/src/main/java/yangFenTuoZi/server/ServerTemplate.java
@@ -1,18 +1,26 @@
 package yangFenTuoZi.server;
 
+import android.annotation.TargetApi;
 import android.app.IActivityManager;
+import android.content.AttributionSource;
+import android.content.Context;
+import android.content.ContextWrapper;
 import android.content.pm.IPackageManager;
 import android.ddm.DdmHandleAppName;
+import android.os.Build;
 import android.os.Handler;
 import android.os.Looper;
 import android.os.RemoteException;
 import android.os.ServiceManager;
 import android.system.Os;
 
-import java.io.File;
-import java.util.Arrays;
+import androidx.annotation.NonNull;
 
-import yangFenTuoZi.server.fakecontext.FakeContext;
+import java.io.File;
+import java.lang.reflect.Constructor;
+import java.lang.reflect.Field;
+import java.lang.reflect.Method;
+import java.util.Arrays;
 
 /**
  * 服务模板类，用于在Android上通过adb/shell..创建一个自己的“服务”
@@ -21,7 +29,7 @@ import yangFenTuoZi.server.fakecontext.FakeContext;
  * @author yangFenTuoZi
  * @version 1.0
  */
-public abstract class ServerTemplate {
+public abstract class ServerTemplate extends ContextWrapper {
     /**
      * 记录器，防止二次崩溃
      */
@@ -35,10 +43,6 @@ public abstract class ServerTemplate {
      * 活动管理器
      */
     public IActivityManager mActivityManager;
-    /**
-     * FakeContext
-     */
-    private FakeContext mContext;
     /**
      * 日志记录器
      */
@@ -62,14 +66,12 @@ public abstract class ServerTemplate {
         public final File logDir;
         public final int[] uids;
         public final boolean enableLogger;
-        public final boolean enableFakeContext;
 
         private Args(Builder builder) {
             serverName = builder.serverName;
             logDir = builder.logDir;
             uids = builder.uids;
             enableLogger = builder.enableLogger;
-            enableFakeContext = builder.enableFakeContext;
         }
 
         public static class Builder {
@@ -77,7 +79,6 @@ public abstract class ServerTemplate {
             public File logDir;
             public int[] uids = new int[0];
             public boolean enableLogger = false;
-            public boolean enableFakeContext = false;
 
             public Builder() {
             }
@@ -93,15 +94,17 @@ public abstract class ServerTemplate {
      * 包括设置主线程、权限检查、日志记录器初始化、异常处理等
      */
     public ServerTemplate(Args args) {
+        super(getSystemContext());
+
         mArgs = args;
         // 切换到主线程
         if (Looper.getMainLooper() == null)
             Looper.prepareMainLooper();
 
         // 判断uid
-        int uid = Os.getuid();
-        if (Arrays.binarySearch(mArgs.uids, uid) == -1) {
-            System.err.printf("Insufficient permission! Need to be launched by %s, but your uid is %d.\n", Arrays.toString(mArgs.uids), uid);
+        UID = Os.getuid();
+        if (Arrays.binarySearch(mArgs.uids, UID) == -1) {
+            System.err.printf("Insufficient permission! Need to be launched by %s, but your uid is %d.\n", Arrays.toString(mArgs.uids), UID);
             System.exit(255);
         }
 
@@ -109,7 +112,7 @@ public abstract class ServerTemplate {
         onCreate();
 
         // 设置程序名称
-        DdmHandleAppName.setAppName(mArgs.serverName, uid);
+        DdmHandleAppName.setAppName(mArgs.serverName, 0);
 
         // 如果启用Logger那么就设置为正常的Logger，否则就设置为空模板Logger
         mLogger = mArgs.enableLogger ? new Logger(mArgs.serverName, mArgs.logDir) : new Logger();
@@ -121,8 +124,8 @@ public abstract class ServerTemplate {
         mPackageManager = IPackageManager.Stub.asInterface(ServiceManager.getService("package"));
         mActivityManager = IActivityManager.Stub.asInterface(ServiceManager.getService("activity"));
 
-        // 看情况创建FakeContext
-        if (mArgs.enableFakeContext) createFakeContext(uid);
+        // 初始化FakeContext
+        initFakeContext();
 
         // 创建一个Handler，为runOnMainThread奠定基础
         mHandler = new Handler();
@@ -136,22 +139,18 @@ public abstract class ServerTemplate {
     }
 
     /**
-     * 创建模拟上下文（FakeContext）
-     *
-     * @param uid 当前用户的UID
+     * 初始化模拟上下文（FakeContext）
      */
-    private void createFakeContext(int uid) {
-        // 准备创建FakeContext
-        mLogger.i("prepare to create FakeContext");
+    private void initFakeContext() {
 
         // 获取当前uid对应的包名
         String packageName;
-        if (uid == 0) packageName = "root";
+        if (UID == 0) packageName = "root";
         else {
             try {
-                packageName = mPackageManager.getPackagesForUid(uid)[0];
+                packageName = mPackageManager.getPackagesForUid(UID)[0];
             } catch (RemoteException e) {
-                mLogger.e("cannot get the package name corresponding to this UID: %d", uid);
+                mLogger.e("cannot get the package name corresponding to this UID: %d", UID);
                 return;
             }
         }
@@ -159,12 +158,10 @@ public abstract class ServerTemplate {
             mLogger.e("got an empty package name");
             return;
         }
-        mLogger.i("create FakeContext { UID = %d, packageName = \"%s\"}", uid, packageName);
+        mLogger.i("init FakeContext { UID = %d, PACKAGE_NAME = \"%s\"}", UID, packageName);
 
-        // 创建FakeContext
-        mContext = new FakeContext();
-        mContext.setPackageName(packageName);
-        mContext.setUid(uid);
+        // 设置FakeContext uid和包名
+        PACKAGE_NAME = packageName;
     }
 
     /**
@@ -242,15 +239,6 @@ public abstract class ServerTemplate {
     }
 
     /**
-     * 获取虚假上下文实例
-     *
-     * @return <code>FakeContext</code>实例
-     */
-    public FakeContext getContext() {
-        return mContext;
-    }
-
-    /**
      * 获取日志记录器实例
      *
      * @return <code>Logger</code>实例
@@ -259,4 +247,71 @@ public abstract class ServerTemplate {
         return mLogger;
     }
 
+    // Workarounds
+
+    private static final Class<?> ACTIVITY_THREAD_CLASS;
+    private static final Object ACTIVITY_THREAD;
+
+    static {
+        if (Looper.getMainLooper() == null)
+            Looper.prepareMainLooper();
+
+        try {
+            ACTIVITY_THREAD_CLASS = Class.forName("android.app.ActivityThread");
+            Constructor<?> activityThreadConstructor = ACTIVITY_THREAD_CLASS.getDeclaredConstructor();
+            activityThreadConstructor.setAccessible(true);
+            ACTIVITY_THREAD = activityThreadConstructor.newInstance();
+
+            Field sCurrentActivityThreadField = ACTIVITY_THREAD_CLASS.getDeclaredField("sCurrentActivityThread");
+            sCurrentActivityThreadField.setAccessible(true);
+            sCurrentActivityThreadField.set(null, ACTIVITY_THREAD);
+        } catch (Exception e) {
+            throw new AssertionError(e);
+        }
+    }
+
+    static Context getSystemContext() {
+        try {
+            Method getSystemContextMethod = ACTIVITY_THREAD_CLASS.getDeclaredMethod("getSystemContext");
+            return (Context) getSystemContextMethod.invoke(ACTIVITY_THREAD);
+        } catch (Throwable throwable) {
+            throwable.printStackTrace(System.err);
+            return null;
+        }
+    }
+    // FakeContext
+
+    private String PACKAGE_NAME;
+    private final int UID;
+
+    @Override
+    public String getPackageName() {
+        return PACKAGE_NAME;
+    }
+
+    @Override
+    @NonNull
+    public String getOpPackageName() {
+        return PACKAGE_NAME;
+    }
+
+    @TargetApi(Build.VERSION_CODES.S)
+    @Override
+    @NonNull
+    public AttributionSource getAttributionSource() {
+        AttributionSource.Builder builder = new AttributionSource.Builder(UID);
+        builder.setPackageName(PACKAGE_NAME);
+        return builder.build();
+    }
+
+    // @Override to be added on SDK upgrade for Android 14
+    @SuppressWarnings("unused")
+    public int getDeviceId() {
+        return 0;
+    }
+
+    @Override
+    public Context getApplicationContext() {
+        return this;
+    }
 }
