--- server/src/main/java/yangFenTuoZi/server/fakecontext/Workarounds.java
@@ -1,44 +1,0 @@
-package yangFenTuoZi.server.fakecontext;
-
-import android.annotation.SuppressLint;
-import android.content.Context;
-import android.os.Looper;
-
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
-
-@SuppressLint("PrivateApi,BlockedPrivateApi,SoonBlockedPrivateApi,DiscouragedPrivateApi")
-public final class Workarounds {
-
-    private static final Class<?> ACTIVITY_THREAD_CLASS;
-    private static final Object ACTIVITY_THREAD;
-
-    static {
-        if (Looper.getMainLooper() == null)
-            Looper.prepareMainLooper();
-
-        try {
-            ACTIVITY_THREAD_CLASS = Class.forName("android.app.ActivityThread");
-            Constructor<?> activityThreadConstructor = ACTIVITY_THREAD_CLASS.getDeclaredConstructor();
-            activityThreadConstructor.setAccessible(true);
-            ACTIVITY_THREAD = activityThreadConstructor.newInstance();
-
-            Field sCurrentActivityThreadField = ACTIVITY_THREAD_CLASS.getDeclaredField("sCurrentActivityThread");
-            sCurrentActivityThreadField.setAccessible(true);
-            sCurrentActivityThreadField.set(null, ACTIVITY_THREAD);
-        } catch (Exception e) {
-            throw new AssertionError(e);
-        }
-    }
-
-    static Context getSystemContext() {
-        try {
-            Method getSystemContextMethod = ACTIVITY_THREAD_CLASS.getDeclaredMethod("getSystemContext");
-            return (Context) getSystemContextMethod.invoke(ACTIVITY_THREAD);
-        } catch (Throwable throwable) {
-            throwable.printStackTrace(System.err);
-            return null;
-        }
-    }
-}
